+++ src/modules/sql.js
@@ -1,26 +1,27 @@
 import sqlParser from '../sqlEngine.js';
-import { TableAlreadyExistsError } from '../errors.js';
+import { InputTypeError, TableAlreadyExistsError } from '../errors.js';
 
 /**
 * SQL module for DataFrame, providing SQL-like syntax for data exploration in DataFrames.
  */
 class SQL {
 
-    static tables = {};
-
     /**
      * Request on a SQL query.
      * @param {String} query A SQL query to request.
      * @returns The result of the query.
-     * @example DataFrame.request('SELECT * FROM tmp');
+     * @example
+     * DataFrame.request('SELECT * FROM tmp');
      */
     static request(query) {
+        if (!(typeof query === 'string')) {throw new InputTypeError('query', ['String'], typeof query);}
         return sqlParser(query, SQL.tables);
     }
 
     /**
      * Drop or remove all registered tables.
-     * @example DataFrame.dropTables();
+     * @example
+     * DataFrame.dropTables();
      */
     static dropTables() {
         SQL.tables = {};
@@ -29,7 +30,8 @@ class SQL {
     /**
      * Drop or remove a registered table.
      * @param {String} tableName The registered table to drop.
-     * @example DataFrame.dropTable('tmp1');
+     * @example
+     * DataFrame.dropTable('tmp1');
      */
     static dropTable(tableName) {
         delete SQL.tables[tableName];
@@ -40,7 +42,8 @@ class SQL {
      * @param {String} tableName The registered table to rename.
      * @param {String} replacement The new table name.
      * @param {Boolean} [overwrite=false] Overwrite if the table already exists.
-     * @example DataFrame.renameTable('tmp1', 'notTmp1');
+     * @example
+     * DataFrame.renameTable('tmp1', 'notTmp1');
      */
     static renameTable(tableName, replacement, overwrite = false) {
         if (SQL.listTables().includes(replacement) && !overwrite) {
@@ -53,7 +56,8 @@ class SQL {
     /**
      * List all registered tables.
      * @returns {Array} A list of the registered tables.
-     * @example DataFrame.listTables();
+     * @example
+     * DataFrame.listTables();
      */
     static listTables() {
         return Object.keys(SQL.tables);
@@ -64,14 +68,11 @@ class SQL {
      * @param {DataFrame} df The DataFrame to register.
      * @param {String} tableName The temporary table name.
      * @param {Boolean} [overwrite=false] Overwrite if the table already exists.
-     * @example DataFrame.registerTable('tmp', df);
+     * @example
+     * DataFrame.registerTable('tmp', df);
      */
     static registerTable(df, tableName, overwrite = false) {
-        if (!(typeof tableName === 'string')) {
-            throw new TypeError(
-                'df must be a DataFrame and tableName a string'
-            );
-        }
+        if (!(df.constructor.name === 'DataFrame')) {throw new InputTypeError('df', ['DataFrame'], df.constructor.name);}
         if (SQL.listTables().includes(tableName) && !overwrite) {
             throw new TableAlreadyExistsError(tableName);
         }
@@ -90,7 +91,8 @@ class SQL {
     /**
      * Register the DataFrame as temporary table.
      * @param {String} tableName The name of the table.
-     * @example df.sql.register('tmp');
+     * @example
+     * df.sql.register('tmp');
      */
     register(tableName) {
         SQL.registerTable(this.df, tableName);
@@ -98,5 +100,8 @@ class SQL {
     }
 
 }
+
+SQL.tables = {};
+
 
 export default SQL;