+++ src/row.js
@@ -38,7 +38,7 @@ class Row {
                 [(value) => (value instanceof Array), () => this._fromArray(data)],
                 [(value) => (value instanceof Row), () => this._fromObject(data[__values__])],
                 [(value) => (typeof value === 'object' && !Object.is(value, null)), () => this._fromObject(data)],
-                [() => true, () => {throw new InputTypeError(typeof data, ['Object', 'Array', 'Row']);}]
+                [() => true, () => {throw new InputTypeError('data', ['Object', 'Array', 'Row'], typeof data);}]
             );
     }
 
@@ -53,6 +53,8 @@ class Row {
     /**
      * Convert Row into dict / hash / object.
      * @returns {Object} The Row converted into dict.
+     * @example
+     * row.toDict()
      */
     toDict() {
         return Object.assign({}, this[__values__]);
@@ -61,6 +63,8 @@ class Row {
     /**
      * Convert Row into Array, loosing column names.
      * @returns {Array} The Row values converted into Array.
+     * @example
+     * row.toArray()
      */
     toArray() {
         return [...this];
@@ -78,6 +82,8 @@ class Row {
      * Check if row contains a column.
      * @param {String} columnName The column to check.
      * @returns {Boolean} The presence or not of the column.
+     * @example
+     * row.has('column1')
      */
     has(columnName) {
         return this[__columns__].includes(columnName);
@@ -87,6 +93,8 @@ class Row {
      * Select columns into the Row.
      * @param {...String} columnNames The columns to select.
      * @returns {Row} A new Row containing only the selected columns.
+     * @example
+     * row.select('column1', 'column2')
      */
     select(...columnNames) {
         return this.__newInstance__(
@@ -100,6 +108,8 @@ class Row {
      * Get a Row value by its column.
      * @param {String} columnToGet The column value to get.
      * @returns The selected value.
+     * @example
+     * row.get('column1')
      */
     get(columnToGet) {
         if (!this.has(columnToGet)) {throw new NoSuchColumnError(columnToGet, this[__columns__]);}
@@ -110,6 +120,8 @@ class Row {
      * Set a Row value by its column, or create a new value if column doesn't exist.
      * @param {String} columnToSet The column value to set.
      * @returns {Row} A new Row with the modified / new value.
+     * @example
+     * row.set('column1', 6)
      */
     set(columnToSet, value) {
         const newRow = Object.assign({}, this[__values__], {[columnToSet]: value});
@@ -120,6 +132,8 @@ class Row {
      * Delete a Row value by its column.
      * @param {String} columnToDel The column value to delete.
      * @returns {Row} A new Row without the deleted value.
+     * @example
+     * row.delete('column1')
      */
     delete(columnToDel) {
         if (!this.has(columnToDel)) {throw new NoSuchColumnError(columnToDel, this[__columns__]);}